Back/requirements.txt
Flask==1.0.2

Flask-SQLAlchemy==2.3.2

Flask-Env==2.0.0
flask-marshmallow==0.9.0

mysqlclient==1.3.13

marshmallow-sqlalchemy==0.14.1

flask-reverse-proxy==0.2.0.2
mysql-connector==2.1.6
mysql-connector-python==8.0.13